move 'keydown' event listener to `Amount` component
